[not verified] rename files to use TS. Update building
