+++ projects/packages/videopress/src/client/block-editor/blocks/video/components/videopress-uploader/index.js
@@ -13,8 +13,8 @@ import { __ } from '@wordpress/i18n';
 import useResumableUploader from '../../../../../hooks/use-resumable-uploader';
 import { uploadFromLibrary } from '../../../../../hooks/use-uploader';
 import { VIDEOPRESS_VIDEO_ALLOWED_MEDIA_TYPES } from '../../constants';
-import { PlaceholderWrapper } from '../../edit.js';
-import { description, title } from '../../index.js';
+import { PlaceholderWrapper } from '../../edit';
+import { description, title } from '../../index';
 import { VideoPressIcon } from '../icons';
 import UploadError from './uploader-error.js';
 import UploadProgress from './uploader-progress.js';
