+++ projects/packages/videopress/src/client/block-editor/blocks/video/components/videopress-uploader/uploader-progress.js
@@ -12,7 +12,7 @@ import filesize from 'filesize';
 import useMetaUpdate from '../../../../../hooks/use-meta-update.js';
 import usePosterImage from '../../../../../hooks/use-poster-image.js';
 import usePosterUpload from '../../../../../hooks/use-poster-upload.js';
-import { PlaceholderWrapper } from '../../edit.js';
+import { PlaceholderWrapper } from '../../edit';
 import UploadingEditor from './uploader-editor.js';
 
 const usePosterAndTitleUpdate = ( { setAttributes, attributes, onDone } ) => {
