+++ projects/packages/videopress/src/client/block-editor/blocks/video/view.ts
